requirements.txt
django-cors-headers==4.3.1
beautifulsoup4==4.12.2
django-crontab==0.7.1
openai==1.5.0
pinecone-client==2.2.4
pypdf2==3.0.1
langchain==0.0.351